Jwt management for incoming requests
Incoming requests are able to decode the Jwt coming in the header field:

Authorization: bearer ...

The Jwt configuration data have been moved within a dedicated section og the appsettings.json.
The IGetLoggedUser service is available and it has been bound to the concrete implementation reading the information from the incoming JWT.
The generated JWT exposes the username in the sub claim (according to RFC7519).