New project: DbFeeder
The DbFeeder project is a console application useful to insert fake data into the database.
It creates 1000 positive subjects and 1000 suspect subjects with all related data and inserts those subjects into the Patients and the Suspects collections. The collections should be empty, since in case unique indexes are active on the collections, case numbers might conflict and an index violation exception might be raised.